+++ kite/src/similarity.rs
@@ -1,26 +1,26 @@
 #[derive(Debug, Clone, PartialEq)]
 pub enum SimilarityModel {
     TfIdf,
-    Bm25{k1: f64, b: f64},
+    Bm25{k1: f32, b: f32},
 }
 
 
 /// tf(term_frequency) = log(term_frequency + 1.0) + 1.0
 #[inline]
-fn tf(term_frequency: u32) -> f64 {
-    (term_frequency as f64 + 1.0f64).ln() + 1.0
+fn tf(term_frequency: u32) -> f32 {
+    (term_frequency as f32 + 1.0f32).ln() + 1.0
 }
 
 
 /// idf(term_docs, total_docs) = log((total_docs + 1.0) / (term_docs + 1.0)) + 1.0
 #[inline]
-fn idf(term_docs: u64, total_docs: u64) -> f64 {
-    ((total_docs as f64 + 1.0) / (term_docs as f64 + 1.0)).ln() + 1.0
+fn idf(term_docs: u64, total_docs: u64) -> f32 {
+    ((total_docs as f32 + 1.0) / (term_docs as f32 + 1.0)).ln() + 1.0
 }
 
 
 impl SimilarityModel {
-    pub fn score(&self, term_frequency: u32, length: f64, total_tokens: u64, total_docs: u64, total_docs_with_term: u64) -> f64 {
+    pub fn score(&self, term_frequency: u32, length: f32, total_tokens: u64, total_docs: u64, total_docs_with_term: u64) -> f32 {
         match *self {
             SimilarityModel::TfIdf => {
                 let tf = tf(term_frequency);
@@ -31,9 +31,9 @@ impl SimilarityModel {
             SimilarityModel::Bm25{k1, b} => {
                 let tf = tf(term_frequency);
                 let idf = idf(total_docs_with_term, total_docs);
-                let average_length = (total_tokens as f64 + 1.0f64) / (total_docs as f64 + 1.0f64);
+                let average_length = (total_tokens as f32 + 1.0f32) / (total_docs as f32 + 1.0f32);
 
-                idf * (k1 + 1.0) * (tf / (tf + (k1 * ((1.0 - b) + b * length.sqrt() / average_length.sqrt())) + 1.0f64))
+                idf * (k1 + 1.0) * (tf / (tf + (k1 * ((1.0 - b) + b * length.sqrt() / average_length.sqrt())) + 1.0f32))
             }
         }
     }
